Support multiple serial streams
Support multiple serial streams by passing the stream reference to the
command handlers. This enables the handler to output errors or success
information to the appropriate stream.

Add support for full matching of the command string.

Change the handler return type from void to boolean so that the command
handlers can indicate that a command has been processed.